+++ ElectronicExam.Administrator/Models/ExamQuestions.cs
@@ -20,7 +20,7 @@ namespace ElectronicExam.Administrator.Models
 
         public string AnsD { get; set; } = "";
 
-        public int Mark { get; set; }
+        public byte Mark { get; set; }
 
         public int ExamId { get; set; }
